+++ source/mt19937ar.cpp
@@ -153,25 +153,25 @@ long genrand_int31(void)
 }
 
 /* generates a random number on [0, 1]-real-interval */
-double genrand_real1(void)
+float genrand_real1(void)
 {
     return genrand_int32()*(1.0/4294967295.0);
 }
 
 /* generates a random number on [0, 1)-real-interval, divided by 2^32 */
-double genrand_real2(void)
+float genrand_real2(void)
 {
     return genrand_int32()*(1.0/4294967296.0);
 }
 
 /* generates a random number on (0, 1)-real-interval */
-double genrand_real3(void)
+float genrand_real3(void)
 {
-    return (((double)genrand_int32()) + 0.5)*(1.0/4294967296.0);
+    return (((float)genrand_int32()) + 0.5)*(1.0/4294967296.0);
 }
 
 /* generates a random number on [0, 1) with 53-bit resolution */
-double genrand_res53(void)
+float genrand_res53(void)
 {
     unsigned long a = genrand_int32()>>5, b = genrand_int32()>>6;
     return (a*67108864.0 + b)*(1.0/9007199254740992.0);
